Switch to using the core/sync xcframework
This is mostly a trivial change that simplifies the build system a bit. There's
only two complicated bits:

Xcode doesn't seem to support configuration-specific target membership, so we
can't just have realm-sync.xcframework be part of the Release configuration and
realm-sync-dbg.xcframework be part of the Debug configuration. Instead,
realm-sync.xcframework is always used, but the path to it is overridden to
actually be pointing at the debug framework for Debug builds. This seems to
work fine and just looks a bit odd in the UI.

Xcode doesn't want to link xcframeworks into static frameworks, but we can just
tell Xcode that the static framework is a regular framework and set the
MACH_O_TYPE to staticlib and then it seems to work fine and do the right thing.